Restructure SBK performance class.
